requirements.txt
streamlit>=1.34
pandas>=2.0
gspread>=6.0
google-auth>=2.0
pyasn1>=0.6.0
rsa>=4.9
